Getting Back Into It
I am working on getting back into coding.  My first steps were to refresh myself on the basics of Microsoft Visual Studio (completed).  Then it was to try and improve some of my old code.  To do that I updated this code to hopefully run faster and lessen the number of screens that pop up.  The final step is to set goals for this code to attempt to complete each week.  Some will be simpler than others, and some will force me to use tools I have not used yet.