Excel Reader Testing
This is the excel reader I have chosen and my thought process of testing it out to see how it works.  This also has the key functions of how the excel reader works.

I am debating on also creating a test of how to pull data based on race before adding it to the main file.